% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fund_imp.R
\name{add_data_loader}
\alias{add_data_loader}
\title{Register a data loader}
\usage{
add_data_loader(fun)
}
\arguments{
\item{fun}{A function to register. Must take no arguments.}
}
\value{
Invisibly returns the updated \code{.fundsr$data_loaders} list.
}
\description{
Appends \code{fun} to the internal data-loader registry (\code{.fundsr$data_loaders}).
Registered functions are intended to be run sequentially in registration
order.
}
\details{
If a loader with the same function body is already registered, \code{fun} is not
added again.
}
\examples{
add_data_loader(function() NULL)
}
