% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{bnpp}
\alias{bnpp}
\title{Import an BNP Paribas fund}
\usage{
bnpp(ticker, file = NULL, benchmark = NULL, session = NULL, ...)
}
\arguments{
\item{ticker}{Fund ticker.}

\item{file}{Optional filename override.}

\item{benchmark}{Optional benchmark key.}

\item{session}{Optional \code{fundsr_session} object. Defaults to the package
default session when \code{NULL}.}

\item{...}{Additional arguments passed to \code{\link[=load_fund]{load_fund()}} and \code{\link[=store_timeseries]{store_timeseries()}}, such as
\code{postprocess}.}
}
\description{
Wrapper around \code{load_fund()} for HSBC files.
}
\seealso{
\code{\link[=load_fund]{load_fund()}}

Other provider wrappers: 
\code{\link{amun}()},
\code{\link{avan}()},
\code{\link{hsbc}()},
\code{\link{inve}()},
\code{\link{ishs}()},
\code{\link{msci}()},
\code{\link{spdr}()},
\code{\link{ubs}()},
\code{\link{vang}()},
\code{\link{xtra}()}
}
\concept{provider wrappers}
