% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fund_imp.R
\name{clear_import_funs}
\alias{clear_import_funs}
\title{Clear registered import functions}
\usage{
clear_import_funs()
}
\value{
Invisibly returns \code{NULL}. Called for side effects.
}
\description{
Clears the internal import-function registry (\code{.fundsr$import_funs}), removing
all previously registered import functions.
}
\examples{
clear_import_funs()
}
