% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{clear_ink_queue}
\alias{clear_ink_queue}
\title{Clear queued Inkscape exports}
\usage{
clear_ink_queue()
}
\value{
Invisibly returns \code{NULL}. Called for side effects.
}
\description{
Clears the internal Inkscape export queue (\code{.fundsr$ink_queue}), removing all
queued export commands.
}
\examples{
clear_ink_queue()
}
