% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_export.R
\name{clear_inkscape_queue}
\alias{clear_inkscape_queue}
\title{Clear queued Inkscape exports}
\usage{
clear_inkscape_queue()
}
\value{
Invisibly returns \code{NULL}. Called for side effects.
}
\description{
Clears the internal Inkscape export queue (\code{.fundsr$inkscape_queue}), removing all
queued export commands.
}
\examples{
clear_inkscape_queue()
}
\seealso{
Other plot export utilities: 
\code{\link{export_pngs}()},
\code{\link{run_plots}()},
\code{\link{save_plot}()}
}
\concept{plot export utilities}
