% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{fundsr_default_session}
\alias{fundsr_default_session}
\title{Get the default fundsr session}
\usage{
fundsr_default_session()
}
\value{
An object of class \code{"fundsr_session"}.
}
\description{
Returns the package-global default \code{fundsr_session} object.
}
\seealso{
Other config functions: 
\code{\link{fundsr_options}()},
\code{\link{fundsr_session}()},
\code{\link{reset_state}()}
}
\concept{config functions}
