% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{fundsr_example_data}
\alias{fundsr_example_data}
\title{Get the path to an example file shipped with the package.}
\usage{
fundsr_example_data(file = ".")
}
\arguments{
\item{file}{The name of the example file.}
}
\description{
Get the path to an example file shipped with the package.
}
\examples{
fundsr_example_data("FNDA.xlsx")
fundsr_example_data()
}
\concept{example helpers}
