% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{get_storage}
\alias{get_storage}
\title{Get the internal fund storage}
\usage{
get_storage(session = NULL)
}
\arguments{
\item{session}{Optional \code{fundsr_session} object. Defaults to the package
default session when \code{NULL}.}
}
\value{
The storage environment.
}
\description{
Returns the fundsr's fund storage environment
(\code{session$storage}).
}
\seealso{
Other fund/index workflow functions: 
\code{\link{add_data_loader}()},
\code{\link{build_all_series}()},
\code{\link{clear_data_loaders}()},
\code{\link{clear_storage}()},
\code{\link{join_env}()},
\code{\link{load_fund}()},
\code{\link{run_data_loaders}()},
\code{\link{store_timeseries}()}
}
\concept{fund/index workflow functions}
