% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{hsbc}
\alias{hsbc}
\title{Import an HSBC fund}
\usage{
hsbc(ticker, file = NULL, benchmark = NULL)
}
\arguments{
\item{ticker}{Fund ticker.}

\item{file}{Optional filename override.}

\item{benchmark}{Optional benchmark key.}
}
\description{
Wrapper around \code{load_fund()} for HSBC files.
}
\seealso{
\code{\link[=load_fund]{load_fund()}}, \code{\link[=amun]{amun()}}, \code{\link[=inve]{inve()}}, \code{\link[=ishs]{ishs()}}, \code{\link[=spdr]{spdr()}}, \code{\link[=ubs]{ubs()}}, \code{\link[=vang]{vang()}}, \code{\link[=xtra]{xtra()}}
}
