% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{inve}
\alias{inve}
\title{Import an Invesco fund}
\usage{
inve(ticker, file = NULL, benchmark = NULL, retrieve_benchmark = FALSE)
}
\arguments{
\item{ticker}{Fund ticker.}

\item{file}{Optional filename override.}

\item{benchmark}{Optional benchmark key.}

\item{retrieve_benchmark}{Logical; also import benchmark column.
#'}
}
\description{
Wrapper around \code{load_fund()} for Invesco files.
}
\seealso{
\code{\link[=load_fund]{load_fund()}}, \code{\link[=amun]{amun()}}, \code{\link[=hsbc]{hsbc()}}, \code{\link[=ishs]{ishs()}}, \code{\link[=spdr]{spdr()}}, \code{\link[=ubs]{ubs()}}, \code{\link[=vang]{vang()}}, \code{\link[=xtra]{xtra()}}
}
