% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{ishs}
\alias{ishs}
\title{Import an iShares fund}
\usage{
ishs(ticker, file = NULL, benchmark = NULL, retrieve_benchmark = FALSE)
}
\arguments{
\item{ticker}{Fund ticker.}

\item{file}{Optional filename override.}

\item{benchmark}{Optional benchmark key.}

\item{retrieve_benchmark}{Logical; also import benchmark column.

#' @seealso
\code{\link[=load_fund]{load_fund()}}, \code{\link[=amun]{amun()}}, \code{\link[=hsbc]{hsbc()}}, \code{\link[=inve]{inve()}}, \code{\link[=spdr]{spdr()}}, \code{\link[=ubs]{ubs()}}, \code{\link[=vang]{vang()}}, \code{\link[=xtra]{xtra()}}}
}
\description{
Wrapper around \code{load_fund()} for iShares files.
}
