% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loaders.R
\name{add_data_loader}
\alias{add_data_loader}
\title{Register a data loader}
\usage{
add_data_loader(fun, session = NULL)
}
\arguments{
\item{fun}{A function to register. Must take no arguments.}

\item{session}{Optional \code{fundsr_session} object. Defaults to the package
default session when \code{NULL}.}
}
\value{
Invisibly returns the updated \code{.fundsr$data_loaders} list.
}
\description{
Appends \code{fun} to the internal data-loader registry (\code{.fundsr$data_loaders}).
Registered functions are intended to be run sequentially in registration
order.
}
\details{
If a loader with the same function body is already registered, \code{fun} is not
added again.
}
\examples{
add_data_loader(function() NULL)
}
\seealso{
Other fund/index workflow functions: 
\code{\link{build_all_series}()},
\code{\link{clear_data_loaders}()},
\code{\link{clear_storage}()},
\code{\link{get_storage}()},
\code{\link{join_env}()},
\code{\link{load_fund}()},
\code{\link{run_data_loaders}()},
\code{\link{store_timeseries}()}
}
\concept{fund/index workflow functions}
