% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider-wrappers.R
\name{amun}
\alias{amun}
\title{Import an Amundi fund}
\usage{
amun(ticker, file = NULL, benchmark = NULL)
}
\arguments{
\item{ticker}{Fund ticker.}

\item{file}{Optional filename override.}

\item{benchmark}{Optional benchmark key.}
}
\description{
Wrapper around \code{load_fund()} for Amundi files.
}
