% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{avan}
\alias{avan}
\title{Import an Avantis fund}
\usage{
avan(ticker, file, benchmark = NULL)
}
\arguments{
\item{ticker}{Fund ticker.}

\item{file}{Filename.}

\item{benchmark}{Optional benchmark key.}
}
\description{
Wrapper around \code{read_timesries()} for Avantis files.
}
\seealso{
\code{\link[=read_timeseries]{read_timeseries()}}

Other provider wrappers: 
\code{\link{amun}()},
\code{\link{hsbc}()},
\code{\link{inve}()},
\code{\link{ishs}()},
\code{\link{msci}()},
\code{\link{spdr}()},
\code{\link{ubs}()},
\code{\link{vang}()},
\code{\link{xtra}()}
}
\concept{provider wrappers}
