% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fund_imp.R
\name{clear_fund_index_map}
\alias{clear_fund_index_map}
\title{Clear fund-index map}
\usage{
clear_fund_index_map()
}
\value{
Invisibly returns \code{NULL}. Called for side effects.
}
\description{
Clears the fund-index map stored in \code{.fundsr$fund_index_map}.
}
\examples{
clear_fund_index_map()
}
