% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fund_index_map.R
\name{clear_fund_index_map}
\alias{clear_fund_index_map}
\title{Clear fund-index map}
\usage{
clear_fund_index_map(session = NULL)
}
\arguments{
\item{session}{Optional \code{fundsr_session} object. Defaults to the package
default session when \code{NULL}.}
}
\value{
Invisibly returns \code{NULL}. Called for side effects.
}
\description{
Clears the fund-index map stored in \code{.fundsr$fund_index_map}.
}
\seealso{
Other fund-index map functions: 
\code{\link{add_fund_index_map}()},
\code{\link{get_fund_index_map}()}
}
\concept{fund-index map functions}
