% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fund_imp.R
\name{get_storage}
\alias{get_storage}
\title{Get the internal fund storage}
\usage{
get_storage()
}
\value{
The storage environment.
}
\description{
Returns the fundsr's fund storage environment
(\code{.fundsr_storage}).
}
