% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fund_imp.R
\name{hsbc}
\alias{hsbc}
\title{Import an HSBC fund}
\usage{
hsbc(ticker, file = NULL, benchmark = NULL)
}
\arguments{
\item{ticker}{Fund ticker.}

\item{file}{Optional filename override.}

\item{benchmark}{Optional benchmark key.}
}
\description{
Wrapper around \code{load_fund()} for HSBC files.
}
