% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{inve}
\alias{inve}
\title{Import an Invesco fund}
\usage{
inve(ticker, file = NULL, benchmark = NULL, retrieve_benchmark = FALSE, ...)
}
\arguments{
\item{ticker}{Fund ticker.}

\item{file}{Optional filename override.}

\item{benchmark}{Optional benchmark key.}

\item{retrieve_benchmark}{Logical; also import benchmark column.}

\item{...}{Additional arguments passed to \code{\link[=load_fund]{load_fund()}} and \code{\link[=store_timeseries]{store_timeseries()}}, such as
\code{postprocess}.}
}
\description{
Wrapper around \code{load_fund()} for Invesco files.
}
\seealso{
\code{\link[=load_fund]{load_fund()}}

Other provider wrappers: 
\code{\link{amun}()},
\code{\link{avan}()},
\code{\link{hsbc}()},
\code{\link{ishs}()},
\code{\link{msci}()},
\code{\link{spdr}()},
\code{\link{ubs}()},
\code{\link{vang}()},
\code{\link{xtra}()}
}
\concept{provider wrappers}
