% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fund_imp.R
\name{ishs}
\alias{ishs}
\title{Import an iShares fund}
\usage{
ishs(ticker, file = NULL, benchmark = NULL, retrieve_benchmark = FALSE)
}
\arguments{
\item{ticker}{Fund ticker.}

\item{file}{Optional filename override.}

\item{benchmark}{Optional benchmark key.}

\item{retrieve_benchmark}{Logical; also import benchmark column.}
}
\description{
Wrapper around \code{load_fund()} for iShares files.
}
