% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{ubs}
\alias{ubs}
\title{Import a UBS fund}
\usage{
ubs(ticker, file = NULL, benchmark = NULL)
}
\arguments{
\item{ticker}{Fund ticker.}

\item{file}{Optional filename override.}

\item{benchmark}{Optional benchmark key.}
}
\description{
Wrapper around \code{load_fund()} for UBS files.
}
\seealso{
\code{\link[=load_fund]{load_fund()}}, \code{\link[=amun]{amun()}}, \code{\link[=hsbc]{hsbc()}}, \code{\link[=inve]{inve()}}, \code{\link[=ishs]{ishs()}}, \code{\link[=spdr]{spdr()}}, \code{\link[=vang]{vang()}}, \code{\link[=xtra]{xtra()}}
}
