% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fund_imp.R
\name{ubs}
\alias{ubs}
\title{Import a UBS fund}
\usage{
ubs(ticker, file = NULL, benchmark = NULL)
}
\arguments{
\item{ticker}{Fund ticker.}

\item{file}{Optional filename override.}

\item{benchmark}{Optional benchmark key.}
}
\description{
Wrapper around \code{load_fund()} for UBS files.
}
