% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fund_imp.R
\name{vang}
\alias{vang}
\title{Import a Vanguard fund}
\usage{
vang(ticker, file = NULL, benchmark = NULL)
}
\arguments{
\item{ticker}{Fund ticker.}

\item{file}{Optional filename override.}

\item{benchmark}{Optional benchmark key.}
}
\description{
Wrapper around \code{fund_imp()} for Vanguard files.
}
