% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{vang}
\alias{vang}
\title{Import a Vanguard fund}
\usage{
vang(ticker, file = NULL, benchmark = NULL)
}
\arguments{
\item{ticker}{Fund ticker.}

\item{file}{Optional filename override.}

\item{benchmark}{Optional benchmark key.}
}
\description{
Wrapper around \code{load_fund()} for Vanguard files.
}
\seealso{
\code{\link[=load_fund]{load_fund()}}

Other provider wrappers: 
\code{\link{amun}()},
\code{\link{hsbc}()},
\code{\link{inve}()},
\code{\link{ishs}()},
\code{\link{msci}()},
\code{\link{spdr}()},
\code{\link{ubs}()},
\code{\link{xtra}()}
}
\concept{provider wrappers}
