% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{xtra}
\alias{xtra}
\title{Import an Xtrackers fund}
\usage{
xtra(ticker, file = NULL, benchmark = NULL, retrieve_benchmark = FALSE)
}
\arguments{
\item{ticker}{Fund ticker.}

\item{file}{Optional filename override.}

\item{benchmark}{Optional benchmark key.}

\item{retrieve_benchmark}{Logical; also import benchmark column.}
}
\description{
Wrapper around \code{load_fund()} for Xtrackers files.
}
\seealso{
\code{\link[=load_fund]{load_fund()}}

Other provider wrappers: 
\code{\link{amun}()},
\code{\link{hsbc}()},
\code{\link{inve}()},
\code{\link{ishs}()},
\code{\link{msci}()},
\code{\link{spdr}()},
\code{\link{ubs}()},
\code{\link{vang}()}
}
\concept{provider wrappers}
